Fix more phpStan errors

+++ Modules/Billing/Http/Traits/Order.php
@@ -13,7 +13,11 @@ use Modules\User\Models\User;
 
 trait Order
 {
-    public function orderSave(int $amount): void
+    /**
+     * @param  float  $amount
+     * @return void
+     */
+    public function orderSave(float $amount): void
     {
         $order = new \Modules\Order\Models\Order();
         Cart::where('user_id', Auth::id())->where('order_id', null)->update(['order_id' => $order->id]);
@@ -22,7 +26,7 @@ trait Order
         $order_data['sub_total'] = Helper::totalCartPrice();
         $order_data['quantity'] = Helper::cartCount();
         $order_data['status'] = 'new';
-        $order_data['total_amount'] = $amount;
+        $order_data['total_amount'] = (int)$amount;
         $order_data['payment_method'] = 'stripe';
         $order_data['payment_status'] = 'paid';
         $order->fill($order_data);
